fix: Change seed argument type from float to int
numpy.random.seed() requires an integer, not a float.
This was causing: TypeError: 'float' object cannot be interpreted as an integer

+++ train_subset.py
@@ -196,7 +196,7 @@ def parse_args(argv):
         default=1e-3,
         help="Learning rate (default: %(default)s)",
     )
-    parser.add_argument("--seed", type=float, default=1, help="Set random seed for reproducibility")
+    parser.add_argument("--seed", type=int, default=1, help="Set random seed for reproducibility")
     parser.add_argument("--save_imgs", action="store_true", help="Save output .npy spectrograms")
     
     # We remove these as they are not needed for this script
